Increase ZKIR test coverage by adding expectations
We are lacking test coverage for the `print-zkir` passes because many of the
test cases use pure circuits, which it will not compile.  The expectations
for these tests are simply `(succeeds)` which means it succeeds by not
running the pass in question.

Change the circuits to be impure by adding a ledger field write, and add
output file expectations.  Copy these to the `print-zkir-v3` tests and
comment out the ones that we do not yet pass.